Fix printf argument types
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@10529 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/allocators/dynarray.cpp
@@ -112,7 +112,7 @@ Status da_set_size(DynArray* da, size_t new_size)
 	{
 		ok = vm::Commit(uintptr_t(end), size_delta_pa);
 		if(!ok)
-			debug_printf(L"Commit failed (%p %d)\n", end, size_delta_pa);
+			debug_printf(L"Commit failed (%p %lld)\n", end, (long long)size_delta_pa);
 	}
 	// shrinking
 	else if(size_delta_pa < 0)
